+++ packages/types/src/interfaces/metadata/types.ts
@@ -1,7 +1,7 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { Bytes, Enum, GenericPortableRegistry, Option, Struct, Text, Type, Vec, bool, u8 } from '@polkadot/types';
+import type { Bytes, Enum, Option, PortableRegistry, Struct, Text, Type, Vec, bool, u8 } from '@polkadot/types';
 import type { Si1Field, Si1LookupTypeId, Si1Type, SiLookupTypeId } from '@polkadot/types/interfaces/scaleInfo';
 
 /** @name ErrorMetadataLatest */
@@ -330,12 +330,6 @@ export interface PalletStorageMetadataV14 extends Struct {
   readonly prefix: Text;
   readonly items: Vec<StorageEntryMetadataV14>;
 }
-
-/** @name PortableRegistry */
-export interface PortableRegistry extends PortableRegistryV14 {}
-
-/** @name PortableRegistryV14 */
-export interface PortableRegistryV14 extends GenericPortableRegistry {}
 
 /** @name PortableType */
 export interface PortableType extends PortableTypeV14 {}